rabbit_ff_registry_factory: Lock around registry initialization
This prevents a potential race which was possible when two processes
attempted to initialize the feature flag registry at the same time. One
process attempting to initialize the registry could take a relatively
long time to read the available feature flags meanwhile another process
could read a more recently changed set of available feature flags.
Locking before reading and writing the flag prevents one update from
clobbering the other.

This race was not possible before the registry used a persistent_term
because we read the version of the registry module as the very first
step when initializing. When we created a new module we checked the
version again and retried the update if it changed in the meantime. That
retry mechanism has been removed since we no longer track the current
version of the registry. Instead we can use a local lock for the same
effect.